refactor aggrergation module
